% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpricialODC.R
\name{EstODC}
\alias{EstODC}
\title{A TestUSO Function}
\usage{
EstODC(X, Y, graph = FALSE)
}
\arguments{
\item{X}{a vector of the sample of size m from the distribution F}

\item{Y}{a vector of the sample of size n from the distribution G}

\item{graph}{plot the empirical ODC or not? Default=FALSE}
}
\value{
The value of empirical ODC at points 0, 1/n, 2/n, ..., (n-1)/n,1
}
\description{
This function allows you to compute the empirical ODC.
}
\examples{
X=rnorm(30,0,1);Y=rnorm(40,1,1);
EstODC(X,Y,graph=TRUE)
}
\keyword{Empirical}
\keyword{ODC}
