% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoF4USO.R
\name{GoF4USO}
\alias{GoF4USO}
\title{A TestUSO Function}
\usage{
GoF4USO(X, Y, alpha = 0.05, graph = FALSE, L = 1000, B = 1000)
}
\arguments{
\item{X}{a vector of the sample of size m from the distribution F}

\item{Y}{a vector of the sample of size n from the distribution G}

\item{alpha}{significance level. Default is 0.05}

\item{graph}{plot or not. Default is FALSE}

\item{L}{number of Monte Carolo samples to compute sample-based critical values. Default is 1000}

\item{B}{number of bootstrapped samples. Default is 1000}
}
\value{
Test statistics, critical values, and conclusion; i.e, reject (if 1) or do not reject (if 0)
}
\description{
This function allows you to test against USO based on X and Y via the sample-based-critical-value approach
}
\examples{
X=rnorm(30,0,1);Y=rnorm(40,1,1);
GoF4USO(X,Y,alpha=0.05,graph=TRUE,L=1000,B=1000)
}
\keyword{Goodness-of-fit}
\keyword{USO}
\keyword{for}
\keyword{test}
