% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSM.R
\name{LSM}
\alias{LSM}
\title{A TestUSO Function}
\usage{
LSM(X, Y, graph = FALSE, ts = FALSE)
}
\arguments{
\item{X}{a vector of the sample of size m from the distribution F}

\item{Y}{a vector of the sample of size n from the distribution G}

\item{graph}{plot the empirical ODC, the LSM of the empirical ODC, and the equal distribution line. Default=FALSE}

\item{ts}{caclcualte the test statistics or not? default=FALSE. If true, return the test statistics when p=1, p=2, p=infinity}
}
\value{
The value of the least star-shaped majorant of the empirical ODC at points 0, 1/n, 2/n, ..., (n-1)/n,1
}
\description{
This function allows you to compute the least star-shape majorant of the empirical ODC.
}
\examples{
X=rnorm(30,0,1);Y=rnorm(40,1,1);
LSM(X,Y,graph=TRUE,ts=TRUE)
}
\keyword{Least}
\keyword{majorant}
\keyword{star-shaped}
